+++ docs/ChangeLog.md
@@ -94,6 +94,7 @@
   - edit: work around a bash-4.4..5.1 bug of `exit` outputting time to stderr of exit context `#D1765` 3de751e e61dbaa
 - util: preserve original traps and restore them on unload `#D1775` `#D1776` `#D1777` 398e404
 - progcomp: support `compopt -o ble/no-default` to suppress default completions `#D1789` 0000000
+- sabbrev: support options `-r` and `--reset` to remove entries `#D1790` 0000000
 
 ## Changes